Update and rename R06MET01J.java to R06_MET01_J.java
