--- R06MET01J.java
@@ -1,11 +1,0 @@
-public static int getAbsAdd(int x, int y) {
-  if (x ==Integer.MIN_VALUE || y == Integer.MIN_VALUE) {
-    throw new IllegalArgumentException();
-  }
-  int absX = Math.abs(x);
-  int absY = Math.abs(y);
-  if (absX > Integer.MAX_VALUE - absY) {
-    throw new IllegalArgumentException();
-  }
-  return absX + absY;
-}
